fix(client): update compatibility for v10.0.2.13 client
This extensive update aligns server network protocol, packet serialization,
and world entry sequences with the ArcheAge v10.0.2.13 client, addressing
numerous client crashes and unexpected behaviors.

Key changes include:
- Overhaul "pish/pisc" variable-length integer encoding to precisely match client implementation.
- Adjust packet structures, field types, and lengths to resolve serialization mismatches.
- Introduce required client initialization packets for character select, world entry,
  and HUD systems (e.g., `SCRepreSentCharacter`, `SCWorldLevelInfo`).
- Remove obsolete packets no longer expected by the v10.0.2.13 client
  (e.g., `SCFactionList`, timed credits/loyalty pushes).
- Update NPC spawning to integrate client-derived `npc_spawners.g.json` data,
  enabling multi-point and roaming area placements.
- Add granular configuration toggles for disabling various world content spawns.
- Optimize network ping behavior to prevent physics time corruption on the client.

+++ AAEmu.Game/Core/Packets/G2C/SCLpManagedPacket.cs
@@ -7,7 +7,8 @@ public class SCLpManagedPacket(uint characterId) : GamePacket(SCOffsets.SCLpMana
 {
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(characterId);
+        // 10.0.2.13: a single i64 "type" (charId).
+        stream.Write((ulong)characterId);
         return stream;
     }
 }
