+++ AAEmu.Game/Core/Packets/G2C/SCResponseUIDataPacket.cs
@@ -8,7 +8,9 @@ public class SCResponseUIDataPacket(uint characterId, ushort uiDataType, string 
 {
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(characterId);
+        // 10.0.2.13 body:
+        // type(i64 charId) | uiDataType(u16) | uiData(length-prefixed string) | size(u32).
+        stream.Write((ulong)characterId);
         stream.Write(uiDataType);
         stream.Write(uiData);
         stream.Write(uiData.Length + 1);
